requirements.txt
# Tested on Python 3.10.x
python-aiml==0.9.*
spacy==3.7.*
nltk==3.8.*
simpful==2.10.*
ultralytics==8.0.*
torch==2.1.*
scikit-learn==1.3.*
pyspellchecker==0.8.3
tensorflow==2.15.*
pillow==10.0.*

# Note: After pip install, run: python -m spacy download en_core_web_md 